refactor(07-Objects 2): create separate file, test and data folders
