+++ 06-Objects/index.js
@@ -5,16 +5,7 @@ import {
   multipleObjects,
 } from "./data/objectCount";
 import { shortPrices } from "./data/priceTransformer";
-import groceryPrices from "./data/05-groceryPrices";
 import { noNull, oneNull, multipleNull, allNull } from "./data/nullDeleter";
-import {
-  noSubObject,
-  oneSubKey,
-  oneSubObjMultKeys,
-  multSubObjOneKey,
-  multSubObjMultKeys,
-  allSubObjMixedKeys,
-} from "./data/objectFlattener";
 
 /**
  * #1: isAnObject
@@ -140,45 +131,6 @@ function priceTransformer(arrayOfObj) {
     transposedObj[newKey] = newPrice;
   }
   return transposedObj;
-}
-
-/**
- * Define the function objectFlattener.
- * objectFlattener accepts a single input, inputObj.
- *
- * When objectFlattener comes across a top-level value that is an object, objectFlattener copies that top-level value's sub key-value pairs to the top-level of inputObj.
- * objectFlattener then deletes the top-level key-value pair. See examples.
- *
- * The inputObj is guaranteed to always have at least one key-value pair.
- * All inputObj values are guaranteed to never be null or undefined.
- *
- * HINT: What can you use to test the data type of a value of inputObj?
- *
- * @category 06 - Objects
- * @function objectFlattener
- * @param {Object} inputObj
- * @returns {Object}
- *
- * @example
- * const noObjVal = { one: 1, two: "string" }
- * objectFlattener(noObjVal) // => { one: 1, two: "string" }
- *
- * const oneObjVal = { one: 1, two: { three: false } }
- * objectFlattener(oneObjVal) // => { one: 1, three: false }
- */
-
-function objectFlattener(inputObj) {
-  for (const key in inputObj) {
-    const currValue = inputObj[key];
-    if (isAnObject(currValue)) {
-      for (const subKey in currValue) {
-        const subValue = currValue[subKey];
-        inputObj[subKey] = subValue;
-      }
-      delete inputObj[key];
-    }
-  }
-  return inputObj;
 }
 
 /**
@@ -341,90 +293,6 @@ describe("06 - Objects", () => {
           (values) => priceValues.indexOf(values) !== -1
         );
         expect(hasAllPriceValues).to.be.true;
-      });
-    });
-  });
-
-  describe("#5: objectFlattener", () => {
-    it("returns an object", () => {
-      expect(objectFlattener(noSubObject)).to.be.an("object");
-      expect(objectFlattener(oneSubKey)).to.be.an("object");
-      expect(objectFlattener(oneSubObjMultKeys)).to.be.an("object");
-      expect(objectFlattener(multSubObjMultKeys)).to.be.an("object");
-    });
-
-    describe("when the input has no object values", () => {
-      it("the untouched input is returned", () => {
-        expect(objectFlattener(noSubObject)).to.eql(noSubObject);
-      });
-    });
-
-    describe("when the input has object values", () => {
-      describe("the returned object has each sub object's key-value pairs", () => {
-        describe("when there is one sub object", () => {
-          it("one key-value pair e.g. { three: { four: 1 }}", () => {
-            const flattenedOne = Object.keys(objectFlattener(oneSubKey));
-            const hasOneSubKey = flattenedOne.indexOf("four") !== -1;
-            expect(hasOneSubKey).to.be.true;
-          });
-
-          it("multiple key-value pairs e.g. { three: { four: 3, five: 'hello' }}", () => {
-            const addedKeys = ["four", "five"];
-            const flattenedOneMult = Object.keys(
-              objectFlattener(oneSubObjMultKeys)
-            );
-            const hasOneSubMultKeys = addedKeys.every(
-              (key) => flattenedOneMult.indexOf(key) !== -1
-            );
-            expect(hasOneSubMultKeys).to.be.true;
-          });
-        });
-
-        describe("when there are multiple sub objects", () => {
-          it("each having one key-value pair e.g. { one: { two: 1 }, three: { four: 4 } }", () => {
-            const addedMultSubObjKeys = ["four", "five"];
-            const flatMultSubOneKeys = Object.keys(
-              objectFlattener(multSubObjOneKey)
-            );
-            const hasMultSubOneKeys = addedMultSubObjKeys.every(
-              (key) => flatMultSubOneKeys.indexOf(key) !== -1
-            );
-
-            expect(hasMultSubOneKeys).to.be.true;
-          });
-
-          it("for multiple sub objects key-value pairs", () => {
-            const multObjMultKeys = ["four", "five", "six", "seven", "eight"];
-            const flatMultObjMultKeys = Object.keys(
-              objectFlattener(multSubObjMultKeys)
-            );
-            const hasMultObjMultKeys = multObjMultKeys.every(
-              (key) => flatMultObjMultKeys.indexOf(key) !== -1
-            );
-            expect(hasMultObjMultKeys).to.be.true;
-          });
-        });
-
-        describe("when all input object values", () => {
-          it("are objects with a mixed number of key-value pairs", () => {
-            const allFlattenedKeys = [
-              "four",
-              "five",
-              "six",
-              "seven",
-              "eight",
-              "nine",
-            ];
-            const flatAllMixedKeys = Object.keys(
-              objectFlattener(allSubObjMixedKeys)
-            );
-            const hasAllMixedKeys = allFlattenedKeys.every(
-              (key) => flatAllMixedKeys.indexOf(key) !== -1
-            );
-
-            expect(hasAllMixedKeys).to.be.true;
-          });
-        });
       });
     });
   });
